refactor(test): move unit tests
