+++ src/test/unit/suite/appc.test.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import { describe, it } from 'mocha';
 
-import Appc from '../../appc';
-import { IosCertificateType } from '../../types/common';
+import Appc from '../../../appc';
+import { IosCertificateType } from '../../../types/common';
 import info from './fixtures/data/ti_info';
 
 Appc.info = info;
