+++ src/test/unit/suite/styleAutoCompleteProvider.test.ts
@@ -6,11 +6,11 @@ import * as path from 'path';
 import * as sinon from 'sinon';
 import * as tce from 'titanium-editor-commons';
 import * as vscode from 'vscode';
-import project from '../../project';
+import project from '../../../project';
 
-import { StyleCompletionItemProvider } from '../../providers/completion/styleCompletionItemProvider';
+import { StyleCompletionItemProvider } from '../../../providers/completion/styleCompletionItemProvider';
 
-const fixturesPath = path.join(__dirname, '../../..', 'src', 'test', 'suite', 'fixtures');
+const fixturesPath = path.join(__dirname, '../../../..', 'src', 'test', 'unit', 'suite', 'fixtures');
 const file = path.join(fixturesPath, 'sample.tss');
 const uri = vscode.Uri.file(file);
 const rawData = fs.readFileSync(path.join(fixturesPath, 'data', 'completions.json'), 'utf8');
